+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog
 
+## 3.5.0
+
+- Support css margin/padding inline/block
+- Bump AngleSharp to 1.5 due to CVE
+
+## 3.4.0
+
+- Numbering list now support `list-style-type: dash`
+- Add support for custom bullet symbols `list-style-type: '👍'` thanks to @AlexAbd1990
+- Minor internal optimisations
+
+## 3.3.2
+
+- Supports Greek numbering in ordered lists (upper-greek / lower-greek) #227
+- Fix nested table in list not filing the full width space #225
+- Fix remove extra whitespace when parsing nested span #224
+
 ## 3.3.1
 
 - Improve parsing to correctly handle spaces in css style #222